backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlmodel
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
starlette
bcrypt==3.2.0
fpdf==1.7.2
